connection.py updated and cleaned the message_handler.py
